removes no-longer-there diskutil verb
As of El Capitan, `sudo diskutil repairPermissions /` returns `diskutil: did not recognize verb "repairPermissions"; type "diskutil" for a list` since repairPermissions has been removed.

In the interest of keeping this tldr-compatible I've removed it. In the interest of not leaving things out, there could instead be a "this no longer works" notice.