remove manual mode and fix missing wildcard
The wildcard went completely missing somewhere before the merge. In the original
design, wildcard certificate usage was about to be referenced with a manual mode example. Running a manual mode with the current example would not produce a wildcard certificate because of the missing \*. subdomain. Furthermore, it would not do anything because a manual DNS mode requires a sort of confirmation command, more info https://github.com/Neilpang/acme.sh/issues/1029. Because of the many quirks of the various DNS modes, a separate subcommand page `acme.sh dns` was created.