circo, fdp, neato, osage, patchwork, sfdp, twopi : add pages
See #6270, #2580.

This pull-request contains the other seven graphviz layout commands in
addition to `dot`, which is awaiting approval at #6270. These files
are generated from the same template as `dot.md` there. The eight
commands differ in the command name, description line and the content
of the `echo` command in one example. (If any more changes are
required, I will regenerate all the files together.)