circo, fdp, neato, osage, patchwork, sfdp, twopi: add page (#6284)
* dot: prepare to add all graphviz layout engines

These changes are in preparation for adding commands: `neato`,
`circo`, `twopi`, `pfd`, `spfd`, `osage` and `patchwork`, which use
the same command syntax. This one is to argue over before posting the
other seven that are generated from the same template. Changes:

- use `gv` extension, not `dot`. Author uses `gv`.
- link to command usage guide page
- specify the type of layout in the description (varies)
- use `input` and `image` not `file` or `out_file`
- provide a list of available image formats
- show how to use the engine as a filter
- give the help command

* dot: add link angle-brackets

* Added a static list of layout engines as line 2 of description.

* dot: Accept suggestion to highlight graph type with back-quotes.

* Fix back-quotes around `png`.

* circo, fdp, neato, osage, patchwork, sfdp, twopi : add pages

See #6270, #2580.

This pull-request contains the other seven graphviz layout commands in
addition to `dot`, which is awaiting approval at #6270. These files
are generated from the same template as `dot.md` there. The eight
commands differ in the command name, description line and the content
of the `echo` command in one example. (If any more changes are
required, I will regenerate all the files together.)

* gv layouts: space after flags. -o versus -O. 80 char 2nd line.

* dot: remove back-quotes from stdin, stdout to match tldr usage.

* graphviz layouts : use https not http.

* graphviz layouts : use {{png}} {{svg}} and {{gif}} as requested.