contributing: explain common contrib issues
Add a clarification that translations are the exception to the single file
per PR rule.

Add a section to explain how to create and accept suggestions within a PR.

Add emphasis to not create PRs from a forked repo's `main` branch and to create
PRs from a named feature branch.

Add emphasis to not `git push -f` on PR branches.

Add images used in `CONTRIBUTING.md` to illustrate the
create and accept suggestions functionality.