added --ignore-missing example
`--ignore-missing` don't fail or report status for missing files when verifying checksums.
This is useful when verifying a subset of downloaded files given a larger list of checksums.
A project could have a checksum for every file in the project but you are only interested in
a few files. This handy option, saves you from extracting the relevant checksums from a file.sha256.

Co-authored-by: Axel Navarro <navarroaxel@gmail.com>