Soop.md : Fix command to update all packages
`Scoop.md` advises to update all packages with `scoop update *`. 

The result of that command is dependent on the shell being used. When using bash, `*` is error-prone as it is expanded to the content of the directory where the command is executed.

`scoop update -a` guarantees consistent results and is supported by scoop's built-in help.

```shell
$ scoop update --help
Usage: scoop update <app> [options]
...
  -a, --all                 Update all apps (alternative to '*')
```