Add TLDR of systemd-cryptenroll
Hello there, I stumbled upon systemd-cryptenroll today and when trying to understand it, I found it had no TLDR yet.

I tried to make it as compact as possible, but the very dynamic parameter set of this command did not allow me to compress it further without leaving out important points. Furthermore, I had to use H2-Headings to distance the different Unlocking Methods from each other. 

I wasn't able to find a relatable / about as complex example to base this one on, but I would be inclined to adapt this pull request if you could provide me with one. 

Also, the Heading/Introduction had to be larger than normal because of the important distinction from the system boot behavior defined in crypttab and to tell the User that the volume needs to be unlocked by systemd-cryptenroll via a currently present Method to make any changes.