CI/codespell: revert changed-files action removal (#11022)
* CI/codespell: revert changed-files action removal

This reverts commit eca9c1e539dc715270105ead81f6f2b59eeec5ee from #10704.

Post this change codespell doesn't show annotations for the PR contents
as it now scanning the whole pages directory and completes execution
after showing a limited number of strings. This effectively made
Codespell unusable. Therefore I have reverted the change to fix
Codespell, the only issue being this would reintroduce the false
detection with annotations being
made in translation PRs on non-English pages.

* CI/codespell: use point release instead of master

Using the development branch isn't ideal for production as it is prone to breakage with a faulty bug or PR. So in this commit, I am updating the codespell action to use the point release instead for more stability and versatility.

Signed-off-by: K.B.Dharun Krishna <kbdharunkrishna@gmail.com>

---------

Signed-off-by: K.B.Dharun Krishna <kbdharunkrishna@gmail.com>
Co-authored-by: Sebastiaan Speck <12570668+sebastiaanspeck@users.noreply.github.com>