expand: fix short option `-t` argument separator
The `expand` command doesn't accept the equal sign '=' as a separator
between the short option `-t` and its arguments. Below are illustrations
of the error the user faces in case the equal sign is used:

  $ expand -t=2 path/to/file
    expand: tab size contains invalid character(s): ‘=2’
  $ expand -t=1,4,6
    expand: tab size contains invalid character(s): ‘=1,4,6’

For the short option `-t`, the correct separator is space, so fix this
in the examples.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>