pages.id: refresh page
This commit brings the upstream changes to the Indonesian translations, including:

+ Refresh outdated pages, e.g. for `node` and `python`
+ Rewrite some pages to use proper Indonesian imperative mood equivalence, part of https://github.com/tldr-pages/tldr/issues/8576
+ Standardize writing style for package manager-related documentation, such as `apt` and `winget`
+ Replace deprecated placeholder names `{{alamat/ke/file_atau_direktori}}` to `{{jalan/menuju/berkas_atau_direktori}}`

Note that the `rubocop` documentation may trigger an "outdated translation" warning. This is false positive as the Indonesian translation updates some placeholders to use the newer [grouped placeholder syntax](https://github.com/tldr-pages/tldr/blob/main/contributing-guides/style-guide.md#grouping-placeholders) which is not applied to the English version yet. The English version must be changed accordingly to reflect the updated guidelines.