chatgpt, libreoffice: move to common and add Indonesian translation
The `chatgpt` command relies heavily on other `common` programs including `bash` and `curl`. This also means the command is available outside of Linux.

Similarly, the `libreoffice` command is also accessible to non-Linux platforms. In macOS, for example, this is located under `/Applications/LibreOffice.app/Contents/MacOS/soffice`. Note that the `soffice` is the actual command name in reference to the Sun StarOffice, the predecessor of LibreOffice (and Apache OpenOffice.org).