Update ffmpeg.md
moved `-ss` and `-to` flags to after input file so they function as output options. using the example provided will result in artifacting before the ffmpeg encounters the first keyframe in the input file. moving them to output options will force ffmpeg to decode the entire input file, which may result in performance overhead, but avoids artifacting in the output file