dos2unix, mac2unix, unix2dos, unix2mac: Standardize More Information links, add Indonesian translation.
This PR encourages the use of manned.org links in place of the original project website over https://waterlan.home.xs4all.nl/dos2unix.html, adding long-form options, and also Indonesian translations.

Note that the "More Information" links of other translations are not replaced as most of them are outdated in terms of content.