chromium, pages/*: redirect common Chromium-based browsers as aliases, add Indonesian translations
There is a lot of Chromium-based browsers and related commands, e.g. Google Chrome which is more known as `google-chrome-stable` on Linux and Microsoft Edge which is known as `msedge` in Windows and `microsoft-edge` in Linux.

I feel the best way to maintain these commands is to redirect them as alias pages.