guides - Update for alias es
In Spanish we are "inviting" the user to take an action to be able to
find documentation for the alias. As terse as before, but different in
spirit.

In the other hand for non aliases we describe what the command is doing,
we will add an style guide for spanish if necessary.