CLIENT-SPECIFICATION: remove `-S` and `-E` (#15795)
I think that these short versions of `--short-options` and
`--long-options` are unintuitive and it might be better to only have the
long variants. If users ask for short variants, we can still add them
later to the specification. On the other hand, releasing them now means
that it will be a breaking change to remove them again.

Co-authored-by: Sebastiaan Speck <12570668+sebastiaanspeck@users.noreply.github.com>