Add account get command for obtaining essential info
It's very important for new users to the mullvad CLI to know this command, both to have a quick reminder of when the account expires, and to obtain the name of the specific computer's autogenerated mullvad 'device name'. This is particularly useful when a user hits 5 devices being used at the same time, but doesn't want to accidentally sign out their currently in-use devices.