Added translation for emerge command (pt_BR)
The emerge command is one of the most important commands in Gentoo Linux, and needed a translation
The command sudo emerge --sync in now deprecated, and should now be used sudo emaint --sync. You can verify this information here: https://wiki.gentoo.org/wiki/Ebuild_repository#Repository_synchronization. 
In the translation, I preserved the deprecated command since is still a compatibility command, but it should eventually be updated in the original file.
I followed the project style and this is my first open source contribution!