rdiff-backup: add page (#17936)
* rdiff-backup: add page

Add a page for `rdiff-backup` (<https://rdiff-backup.net>). There's only
a minimal set of examples -- one about backing up, one about listing
backups, and two related to restoring -- but this can serve as a
starting point for more.

* Update pages/common/rdiff-backup.md

change "more information" URL

Co-authored-by: Managor <42655600+Managor@users.noreply.github.com>

* Update pages/common/rdiff-backup.md

Paths to a repository don't need to be absolute, change the examples to make that a bit more clear.

Co-authored-by: Managor <42655600+Managor@users.noreply.github.com>

---------

Co-authored-by: Managor <42655600+Managor@users.noreply.github.com>